+++ iterm/term_notes.py
@@ -29,7 +29,7 @@ import tomllib
 
 import iterm2
 
-__version__ = "0.3.1"
+__version__ = "0.3.2"
 
 CONFIG_PATH = os.path.expanduser(
     os.environ.get("TERM_NOTES_CONFIG", "~/.config/term-notes/config.toml"))